Move Navbar to index
